Final set of changes based on feedback
After further investigation it was found that the Makefile was missing
calls to CFLAGS which was causing issues with notes generation. That has
now been fixed.

Also the CFLAGS and LDFLAGS have been updated to match the latest best
values provided to us by release-delivery.

Additionally the CFLAGS and LDFLAGS configuration have been moved out of
the external build_checks.sh script into the Makefile directly so that
the build_checks script can just run make directly, as well as setting
proper defaults for how we expect sg-bridge to be built.

Signed-off-by: Leif Madsen <lmadsen@redhat.com>