Fix and test Dockerfile builds
When making fixes in #29 related to lint checks, a local build wasn't
done against the Dockerfile, which has a different base image from that
used by build_checks for lint verification. The result of that was that
building sg-bridge with the Dockerfile failed on gcc compiling due to
hardening instructions not being available.

This commit does the following:

* adds the RPM dependency to the build layer of Dockerfile to install
  hardening instructions for gcc
* changes the default -march CFLAG to be x86-64 instead of x86-64-v2
  which fails in ubi8
* updates the GHA to add a build check against the Dockerfile and not
  just run build_checks.sh for future validation

Found by csibbitt and noted in chat.

Resolves #30

Signed-off-by: Leif Madsen <lmadsen@redhat.com>